Crate Myproject

MyProject contain 5 modules.

Modules

area

The area module contains values and function calling.

common_struct

The common_struct module contains common data structures used in the crate.

employee

The employee module contains functionality related to employees.

employee_hashmap

The employee module contains functionality related to employees.

student

The student module contains functionality related to students.

student_hashmap

The student module contains functionality related to students.

table_task

The table_task module contains functionality related to table tasks.

table_task_hashmap

The table_task module contains functionality related to table tasks. and

two_string

The two_string module contains functionality related to operations on two strings.

Functions

main

The main entry point for the application.